llama3/requirements.txt
langchain_community.llms
streamlit
langchain_core.prompts